+++ camel-ladok3-component/src/main/java/se/kth/infosys/smx/ladok3/internal/Ladok3StudiedeltagandeServiceWrapper.java
@@ -1,5 +1,7 @@
 package se.kth.infosys.smx.ladok3.internal;
 
+import java.util.HashMap;
+import java.util.Iterator;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -15,19 +17,23 @@ import se.kth.infosys.ladok3.StudiedeltagandeServiceImpl;
 import se.kth.infosys.smx.ladok3.Ladok3Message;
 import se.ladok.schemas.studentinformation.Student;
 import se.ladok.schemas.studiedeltagande.IngaendeKurspaketeringstillfalleLista;
+import se.ladok.schemas.studiedeltagande.StudieaktivitetUtdata;
 import se.ladok.schemas.studiedeltagande.TillfallesdeltagandeLista;
 
 public class Ladok3StudiedeltagandeServiceWrapper implements Ladok3ServiceWrapper {
     private static final Logger log = LoggerFactory.getLogger(Ladok3StudiedeltagandeServiceWrapper.class);
     private static final Pattern URL_PATTERN = Pattern.compile(
             "^/studiedeltagande(/(?<operation>"
-            + "pabarjadtbildning/kurspaketering/student|"
+            + "tillfallesdeltagande/kurstillfallesdeltagande/student|"
+            + "utdata/studieaktivitetochfinansiering|"
+            + "pabarjadutbildning/kurspaketering/student|"
             + "studiestruktur/student"
             + "))+.*");
     private StudiedeltagandeService service;
     private String pathOperation;
 
-    public Ladok3StudiedeltagandeServiceWrapper(String host, String path, SSLContext context) throws Exception {
+    public Ladok3StudiedeltagandeServiceWrapper(String host, String path, SSLContext context)
+            throws Exception {
         this.service = new StudiedeltagandeServiceImpl(host, context);
         Matcher matcher = URL_PATTERN.matcher(path);
         if (matcher.matches()) {
@@ -38,7 +44,13 @@ public class Ladok3StudiedeltagandeServiceWrapper implements Ladok3ServiceWrappe
     public void doExchange(Exchange exchange) throws Exception {
         final String operation = currentOperation(exchange);
         switch (operation) {
-        case "pabarjadtbildning/kurspaketering/student":
+        case "tillfallesdeltagande/kurstillfallesdeltagande/student":
+            handleKurstillfallesdeltagandeStudent(exchange);
+            break;
+        case "utdata/studieaktivitetochfinansiering":
+            handleStudieaktivitetOchFinansiering(exchange);
+            break;
+        case "pabarjadutbildning/kurspaketering/student":
             handlePabarjadutbildningKurspaketeringStudent(exchange);
             break;
         case "studiestruktur/student":
@@ -47,6 +59,30 @@ public class Ladok3StudiedeltagandeServiceWrapper implements Ladok3ServiceWrappe
         default:
             throw new CamelExchangeException("Unupported operation: %s" + operation, exchange);
         }
+    }
+
+    private void handleKurstillfallesdeltagandeStudent(Exchange exchange) throws Exception {
+        String uid = exchange.getIn().getHeader(Ladok3Message.Header.KeyValue, String.class);
+        if (uid == null || uid.isEmpty()) {
+            Student student = exchange.getIn().getMandatoryBody(Student.class);
+            uid = student.getUid();
+        }
+
+        log.debug("Getting kurstillfallesdeltaganden for student with uid: {}", uid);
+        TillfallesdeltagandeLista fromLadok = service.kurstillfallesdeltagandeStudent(uid);
+        exchange.getIn().setBody(fromLadok);
+    }
+
+    private void handleStudieaktivitetOchFinansiering(Exchange exchange) {
+        @SuppressWarnings("unchecked")
+        HashMap<String, Object> params = exchange.getIn().getHeader(
+                Ladok3Message.Header.Params, new HashMap<String, Object>(), HashMap.class);
+
+        log.debug("Getting Ladok data for studieaktivetet och finansiering request with params: {}",
+                params);
+        Iterator<StudieaktivitetUtdata> fromLadok =
+                service.utdataStudieaktivitetOchFinansieringIteraterable(params).iterator();
+        exchange.getIn().setBody(fromLadok);
     }
 
     private void handleStudiestrukturStudent(Exchange exchange) throws Exception {